Extract security section in settings
Related to #1132
