+++ apps/mobile-wallet/src/navigation/RootStackNavigation.tsx
@@ -14,7 +14,7 @@ import FundPasswordScreen from '~/features/fund-password/FundPasswordScreen'
 import { deleteFundPassword } from '~/features/fund-password/fundPasswordStorage'
 import AppModals from '~/features/modals/AppModals'
 import { loadBiometricsSettings } from '~/features/settings/settingsPersistentStorage'
-import SettingsScreen from '~/features/settings/SettingsScreen'
+import SettingsScreen from '~/features/settings/settingsScreen/SettingsScreen'
 import { useAppDispatch, useAppSelector } from '~/hooks/redux'
 import { useBiometricsAuthGuard } from '~/hooks/useBiometrics'
 import AlephiumLogo from '~/images/logos/AlephiumLogo'
